License project under LPPL v1.3c
Since the code is being shared publicly and up to now it should have
specified a license in order to allow other students to use and modify
the code for their own theses.  Note that LPPL covers only the LaTeX
code and packages, not the actual contents of the generated document.
The thesis contents still remain subject of copyright law.