Fix statutory declaration (issues #9 and #11)
The statutory declaration section was fixed according official
guidelines.

 - Refactored original template code related to "Statutory declaration"
   section
 - Moved the section to the right position in thesis.
 - Changed text of the declaration.

According to the official guidelines, the thesis title should be
handwritten, nevertheless it's far more elegant to typeset it in italic.
The official guidelines also do not state there should be a place, date
next to the subscription, but I think that it should be there, as well
as the actual student's name below the dotted line.