Add monkey patch for \cmidrule command
The slovak babel package redefines '-' as an active character for whole
code.  This behavior interferes with all commands using plaintext '-'
character in their arguments, such as \cline and it's derivatives.
Nevertheless, some tables in this thesis should use \cmidrule due to
multi-level headers (see Issue #32).

This problem can be cleanly solved by monkey-patching \cmidrule command
to use the active '-' as a separator in it's argument (instead of
plaintext '-').