Replace thin space with regular space
The 0.16667em wide thin space is too narrow to be typeset between
abbreviation and the next word.  Since in Slovak language is using the
French spacing, regular space is approx. 0.33333em wide and thus it is
close enough to the optimal 0.25em.  In tables there is no need for
unbreakable space, because there is no text flow.  I have decided
to use regular spaces there to keep code more readable.