Add better formatting to footnote code
Long footnotes are a kind of pain in the ass in case if they do not
contain enough spaces for filling the paragraph.  Since the \footnote
command itself represents the placeholder for the number in the document
and this number should not be separated by space, the \footnote commands
can make paragraph code quite messy, especially when they get stuck with
the word from previous line.

Luckily there is a nice hack borrowed from TeX.  For instance the
problem shown on the example below:

  a couple of
  words\footnote{\url{https://extremely.long/url/that does-not-fit}}

can be solved by adding % character behind last word and moving footnote
to the next line:

  a couple of words%
  \footnote{\url{https://extremely.long/url/that does-not-fit}}

Both example above give exactly the same result, but the latter code
offers better readability.